move cookie functions to ntske package
